Fix bug in owoga ripgrep
Not sure how this was working before without the mapcar
'shell-quote-argument. Maybe it worked on Linux but not on Mac due to
how each handles shell expansion? The problem came if I tried to use
the arguments ("-g" "*test*"). The asterisks caused something weird to
happen and it would try treating the regex as the directory.

Also fixes indentation.